* PlaylistView.m           use NSDragOperation, and some int -> NSInteger transitions         * Playlist.[h|m]           return NSUInteger for numberOfSongs instead of unsigned int           use NSUInteger with songAtIndex:           return NSUInteger for indexOfSong, and use NSUInteger for           moveSongsAtIndexes:toIndex:         * PlaylistController.m           fix format string after numberOfSongs, and minor int -> NSInteger           transitions         * Frameworks/Cynthiune/utils.h           fix format string
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@2645 d25b4930-3292-e311-b074-0010c6de5da8

+++ user-apps/Cynthiune/Frameworks/Cynthiune/NSArrayExtensions.h
@@ -42,8 +42,8 @@
 
 - (void) addObjectsFromArray: (NSArray*) otherArray
                      atIndex: (unsigned int) index;
-- (unsigned int) moveObjectsAtIndexes: (NSArray *) indexes
-                              toIndex: (unsigned int) index;
+- (NSUInteger) moveObjectsAtIndexes: (NSArray *) indexes
+                              toIndex: (NSUInteger) index;
 - (void) addObjectRandomly: (id) object;
 
 - (void) rotateUpToObject: (id) object;
